Removed use of banks/bankstreams for working with the buffers.  It's just raw memory shagging now.  Also introduced the RENDER_BUFFER_SCALE constant so that you can possibly tweak the amount of memory that gets added to buffers when they have to be resized to accomodate more data.
AddVerticesEx has a new parameter: elements, which comes first in the parameter list.  Arrays are no longer required for copying data to the buffers, only pointers.  Elements specifies the number of vertices described in the buffers.  The color and texcoord buffers are optional (pass Null).  The color buffer will be filled with the value 255 for the range of vertices described, the texcoord buffer will remain untouched (garbage data).

GL_EXT_compiled_vertex_array is used on systems that support it now.  Check for glMultiDrawArrays replaced with a check for GL_VERSION_1_4 - if a system doesn't pass it, the renderer falls back to DrawArrays as before.